Fix __entry address holder type from uint16_t to uint64_t

+++ src/front/cpu.cpp
@@ -92,7 +92,7 @@ int main(int argc, char* argv[]) {
     CPU cpu;
 
     map<string, uint64_t> function_address_mapping = loader.getFunctionAddresses();
-    uint16_t starting_instruction = function_address_mapping["__entry"];
+    uint64_t starting_instruction = function_address_mapping["__entry"];
     for (auto p : function_address_mapping) { cpu.mapfunction(p.first, p.second); }
     for (auto p : loader.getBlockAddresses()) { cpu.mapblock(p.first, p.second); }
